@@ -3,65 +3,48 @@
 	"version": "0.2.0",
 	"license": "AGPL-3.0-or-later",
 	"scripts": {
-		"serve": "vue-cli-service serve",
-		"build": "vue-cli-service build",
-		"lint": "tsc --noEmit && vue-cli-service lint .",
+		"serve": "vite serve",
+		"build": "vite build",
+		"lint": "tsc --noEmit && eslint --ext .js,.ts,.vue --fix .",
 		"i18n:report": "vue-cli-service i18n:report --src \"./src/**/*.?(js|vue)\" --locales \"./src/locales/**/*.json\"",
-		"lint-ci": "tsc --noEmit && vue-cli-service --no-fix lint .",
-		"test": "vue-cli-service test:unit --config ./tests/unit/jest.config.js -experimental-vm-modules --runInBand --detectOpenHandles --forceExit"
+		"lint-ci": "tsc --noEmit && eslint .",
+		"test": "vitest run --coverage",
+		"cy:open": "cypress open",
+		"cy:run": "cypress run --component"
 	},
 	"dependencies": {
 		"@fortawesome/fontawesome-free": "^5.12.1",
 		"@mdi/font": "^3.9.97",
 		"@vimeo/player": "^2.10.0",
-		"@vue/cli": "^5.0.8",
-		"@vue/composition-api": "1.7.0",
+		"@vueuse/core": "^9.6.0",
 		"load-script": "^1.0.0",
 		"material-design-icons-iconfont": "^5.0.1",
+		"ott-common": "./common",
+		"sortablejs": "^1.15.0",
+		"sortablejs-vue3": "^1.2.3",
 		"video.js": "^7.15.4",
-		"vue": "^2.6.14",
+		"vue": "3.2.45",
 		"vue-axios": "^2.1.5",
-		"vue-fragment": "1.5.1",
 		"vue-gtag": "^1.6.1",
-		"vue-i18n": "^8.26.3",
-		"vue-router": "^3.5.1",
-		"vue-slider-component": "^3.1.1",
-		"vuedraggable": "^2.23.2",
-		"vuetify": "^2.6.12",
-		"vuex": "^3.6.2"
+		"vue-i18n": "9.2.2",
+		"vue-router": "^4.1.5",
+		"vue-slider-component": "4.1.0-beta.6",
+		"vuetify": "3.1.2",
+		"vuex": "4.1.0"
 	},
 	"devDependencies": {
-		"@babel/core": "7.19.6",
-		"@cypress/vue": "^2.2.3",
-		"@cypress/webpack-dev-server": "^1.3.1",
-		"@intlify/vue-i18n-loader": "^1.1.0",
+		"@cypress/vue": "^5.0.3",
 		"@types/video.js": "^7.3.29",
 		"@types/vimeo__player": "^2.16.0",
-		"@types/web": "0.0.36",
-		"@vue/cli-plugin-babel": "^5.0.8",
-		"@vue/cli-plugin-e2e-cypress": "^5.0.8",
-		"@vue/cli-plugin-eslint": "^5.0.8",
-		"@vue/cli-plugin-typescript": "^5.0.8",
-		"@vue/cli-plugin-unit-jest": "^5.0.8",
-		"@vue/cli-service": "^5.0.8",
+		"@types/web": "0.0.80",
+		"@vitejs/plugin-vue": "^4.0.0",
+		"@vitest/coverage-c8": "^0.25.1",
 		"@vue/eslint-config-typescript": "^7.0.0",
-		"@vue/test-utils": "1.3.0",
-		"@vue/vue2-jest": "^27.0.0-alpha.4",
-		"babel-eslint": "10.0.3",
-		"babel-jest": "27.5.1",
-		"core-js": "3.6.4",
-		"css-loader": "^5.2.5",
-		"eslint-plugin-vue": "6.1.2",
-		"fork-ts-checker-webpack-plugin": "^6.5.0",
+		"@vue/test-utils": "2.2.1",
+		"eslint-plugin-vue": "9.7.0",
 		"sass": "^1.41.1",
-		"sass-loader": "^10.0.0",
-		"style-loader": "^2.0.0",
-		"ts-loader": "^9.2.6",
-		"vue-cli-plugin-i18n": "~2.3.1",
-		"vue-cli-plugin-vuetify": "^2.5.8",
-		"vue-cli-plugin-webpack-bundle-analyzer": "~4.0.0",
-		"vue-template-compiler": "^2.6.14",
-		"vuetify-loader": "1.7.2",
-		"webpack-bundle-analyzer": "3.6.0"
+		"vite": "^4.0.4",
+		"vite-plugin-vuetify": "1.0.0",
+		"vitest": "^0.25.1"
 	}
 }
