@@ -1,45 +1,40 @@
 <template>
 	<v-app id="app">
-		<v-app-bar
-			app
-			:absolute="!$store.state.fullscreen"
-			:inverted-scroll="$store.state.fullscreen"
-		>
-			<v-app-bar-nav-icon @click="drawer = true" />
+		<v-app-bar app :absolute="!fullscreen" v-if="!fullscreen">
+			<!-- TODO: replace v-if here with hide on scroll and inverted scroll when vuetify 3.2 comes out. -->
+			<v-app-bar-nav-icon @click="drawer = true" role="menu" aria-label="nav menu" />
 			<v-img
-				:src="require('@/assets/logo.svg')"
+				:src="logoUrl"
 				max-width="32"
 				max-height="32"
 				contain
 				style="margin-right: 8px"
 			/>
-			<v-toolbar-title>
-				<router-link class="link-invis" style="margin-right: 10px" to="/">
-					OpenTogetherTube
-				</router-link>
-			</v-toolbar-title>
-			<v-toolbar-items v-if="$vuetify.breakpoint.lgAndUp">
-				<v-btn text to="/rooms">{{ $t("nav.browse") }}</v-btn>
-				<v-btn text to="/faq">{{ $t("nav.faq") }}</v-btn>
+			<v-app-bar-title class="app-bar-title">
+				<router-link class="link-invis" to="/">OpenTogetherTube</router-link>
+			</v-app-bar-title>
+			<v-toolbar-items v-if="$vuetify.display.lgAndUp">
+				<v-btn variant="text" to="/rooms">{{ $t("nav.browse") }}</v-btn>
+				<v-btn variant="text" to="/faq">{{ $t("nav.faq") }}</v-btn>
 				<v-btn
-					text
+					variant="text"
 					href="https://github.com/dyc3/opentogethertube/issues/new/choose"
 					target="_blank"
 				>
-					<v-icon class="side-pad">fas fa-bug</v-icon>
+					<v-icon class="side-pad">fa:fas fa-bug</v-icon>
 					{{ $t("nav.bug") }}
 				</v-btn>
-				<v-btn text href="https://github.com/sponsors/dyc3" target="_blank">
-					<v-icon class="side-pad">fas fa-heart</v-icon>
+				<v-btn variant="text" href="https://github.com/sponsors/dyc3" target="_blank">
+					<v-icon class="side-pad">fa:fas fa-heart</v-icon>
 					{{ $t("nav.support") }}
 				</v-btn>
 			</v-toolbar-items>
 			<v-spacer />
-			<v-toolbar-items v-if="$vuetify.breakpoint.lgAndUp">
+			<v-toolbar-items v-if="$vuetify.display.lgAndUp">
 				<v-menu offset-y>
-					<template v-slot:activator="{ on }">
-						<v-btn text v-on="on">
-							<v-icon class="side-pad">fas fa-plus-square</v-icon>
+					<template v-slot:activator="{ props }">
+						<v-btn variant="text" v-bind="props">
+							<v-icon class="side-pad">fa:fas fa-plus-square</v-icon>
 							{{ $t("nav.create.title") }}
 						</v-btn>
 					</template>
@@ -51,66 +46,41 @@
 					</v-list>
 				</v-menu>
 				<NavUser @login="showLogin = true" @logout="logout" />
-				<v-select
-					solo
-					flat
-					style="margin-top: 5px; width: 100px"
-					:items="locales"
-					@change="setLocale"
-					:value="$i18n.locale"
-				/>
+				<LocaleSelector style="margin-top: 5px; width: 100px" />
 			</v-toolbar-items>
 		</v-app-bar>
 		<v-navigation-drawer v-model="drawer" absolute temporary>
 			<v-list nav dense>
-				<v-list-item-group>
-					<v-list-item to="/">
-						<v-list-item-content>
-							{{ $t("nav.home") }}
-						</v-list-item-content>
-					</v-list-item>
-					<v-list-item to="/rooms">
-						<v-list-item-content>
-							{{ $t("nav.browse") }}
-						</v-list-item-content>
-					</v-list-item>
-					<v-list-item to="/faq">
-						<v-list-item-content>
-							{{ $t("nav.faq") }}
-						</v-list-item-content>
-					</v-list-item>
-					<v-list-item
-						href="https://github.com/dyc3/opentogethertube/issues/new/choose"
-						target="_blank"
-					>
-						<v-list-item-icon>
-							<v-icon>fas fa-bug</v-icon>
-						</v-list-item-icon>
-						<v-list-item-content>
-							{{ $t("nav.bug") }}
-						</v-list-item-content>
-					</v-list-item>
-					<v-list-item href="https://github.com/sponsors/dyc3" target="_blank">
-						<v-list-item-icon>
-							<v-icon>fas fa-heart</v-icon>
-						</v-list-item-icon>
-						<v-list-item-content>
-							{{ $t("nav.support") }}
-						</v-list-item-content>
-					</v-list-item>
-					<NavCreateRoom
-						@createtemp="createTempRoom"
-						@createperm="showCreateRoomForm = true"
-					/>
-					<NavUser @login="showLogin = true" @logout="logout" />
-					<v-select
-						solo
-						flat
-						:items="locales"
-						@change="setLocale"
-						:value="$i18n.locale"
-					/>
-				</v-list-item-group>
+				<v-list-item to="/">
+					{{ $t("nav.home") }}
+				</v-list-item>
+				<v-list-item to="/rooms">
+					{{ $t("nav.browse") }}
+				</v-list-item>
+				<v-list-item to="/faq">
+					{{ $t("nav.faq") }}
+				</v-list-item>
+				<v-list-item
+					href="https://github.com/dyc3/opentogethertube/issues/new/choose"
+					target="_blank"
+				>
+					<template #prepend>
+						<v-icon>fa:fas fa-bug</v-icon>
+					</template>
+					{{ $t("nav.bug") }}
+				</v-list-item>
+				<v-list-item href="https://github.com/sponsors/dyc3" target="_blank">
+					<template #prepend>
+						<v-icon>fa:fas fa-heart</v-icon>
+					</template>
+					{{ $t("nav.support") }}
+				</v-list-item>
+				<NavCreateRoom
+					@createtemp="createTempRoom"
+					@createperm="showCreateRoomForm = true"
+				/>
+				<NavUser @login="showLogin = true" @logout="logout" />
+				<LocaleSelector />
 			</v-list>
 			<template v-slot:append>
 				<div class="pa-2">
@@ -134,20 +104,15 @@
 				<LogInForm @shouldClose="showLogin = false" />
 			</v-dialog>
 		</v-container>
-		<v-overlay :value="createRoomState.value.isLoadingCreateRoom">
-			<v-container fill-height>
-				<v-row align="center" justify="center">
-					<v-col cols="12" sm="4">
-						<v-progress-circular indeterminate />
-						<v-btn
-							elevation="12"
-							x-large
-							@click="cancelRoom"
-							style="margin-top: 24px"
-							>{{ $t("actions.cancel") }}</v-btn
-						>
-					</v-col>
-				</v-row>
+		<v-overlay
+			class="overlay-loading-create-room"
+			:model-value="store.state.misc.isLoadingCreateRoom"
+		>
+			<v-container class="overlay-loading-create-room">
+				<v-progress-circular indeterminate />
+				<v-btn elevation="12" size="x-large" @click="cancelRoom" style="margin-top: 24px">
+					{{ $t("actions.cancel") }}
+				</v-btn>
 			</v-container>
 		</v-overlay>
 		<Notifier />
@@ -155,7 +120,7 @@
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import { defineComponent, onMounted, ref, computed } from "vue";
 import { API } from "@/common-http.js";
 import CreateRoomForm from "@/components/CreateRoomForm.vue";
 import LogInForm from "@/components/LogInForm.vue";
@@ -163,9 +128,13 @@ import NavUser from "@/components/navbar/NavUser.vue";
 import NavCreateRoom from "@/components/navbar/NavCreateRoom.vue";
 import Notifier from "@/components/Notifier.vue";
 import { loadLanguageAsync } from "@/i18n";
-import { createRoomHelper, createRoomState } from "@/util/roomcreator";
+import { createRoomHelper } from "@/util/roomcreator";
+import { useRouter } from "vue-router";
+import logoUrl from "@/assets/logo.svg";
+import { useStore } from "@/store";
+import { LocaleSelector } from "@/components/navbar/LocaleSelector.vue";
 
-export default Vue.extend({
+export const App = defineComponent({
 	name: "app",
 	components: {
 		CreateRoomForm,
@@ -173,87 +142,91 @@ export default Vue.extend({
 		NavUser,
 		NavCreateRoom,
 		Notifier,
+		LocaleSelector,
 	},
-	data() {
-		return {
-			createRoomState,
-			showCreateRoomForm: false,
-			showLogin: false,
-			drawer: false,
+	setup() {
+		const store = useStore();
 
-			locales: [
-				{
-					text: "🇺🇸",
-					value: "en",
-				},
-				{
-					text: "🇩🇪",
-					value: "de",
-				},
-				{
-					text: "🇷🇺",
-					value: "ru",
-				},
-			],
+		const showCreateRoomForm = ref(false);
+		const showLogin = ref(false);
+		const drawer = ref(false);
+
+		const logout = async () => {
+			let res = await API.post("/user/logout");
+			if (res.data.success) {
+				store.commit("LOGOUT");
+			}
 		};
-	},
-	methods: {
-		logout() {
-			API.post("/user/logout").then(res => {
-				if (res.data.success) {
-					this.$store.commit("LOGOUT");
-				}
-			});
-		},
-		async setLocale(locale: string) {
+
+		const setLocale = async (locale: string) => {
 			await loadLanguageAsync(locale);
-			this.$store.commit("settings/UPDATE", { locale });
-		},
-		cancelRoom() {
-			createRoomState.value.cancelledRoomCreation = true;
-			createRoomState.value.isLoadingCreateRoom = false;
-		},
-		async createTempRoom() {
-			await createRoomHelper(this.$store);
-		},
-	},
-	async created() {
-		this.$store.subscribe((mutation, state) => {
-			if (mutation.type === "misc/ROOM_CREATED") {
-				try {
-					// @ts-expect-error because vue router doesn't quite work with ts like this and im feeling lazy.
-					this.$router.push(`/room/${mutation.payload.name}`);
-				} catch (e) {
-					if (e.name !== "NavigationDuplicated") {
-						throw e;
+			store.commit("settings/UPDATE", { locale });
+		};
+
+		const cancelRoom = () => {
+			store.commit("misc/CANCELLED_ROOM_CREATION");
+		};
+
+		const createTempRoom = async () => {
+			await createRoomHelper(store);
+		};
+
+		onMounted(async () => {
+			const router = useRouter();
+
+			store.subscribe(mutation => {
+				if (mutation.type === "misc/ROOM_CREATED") {
+					try {
+						router.push(`/room/${mutation.payload.name}`);
+					} catch (e) {
+						if (e.name !== "NavigationDuplicated") {
+							throw e;
+						}
 					}
 				}
+			});
+
+			document.addEventListener("fullscreenchange", () => {
+				if (document.fullscreenElement) {
+					store.commit("SET_FULLSCREEN", true);
+					document.querySelector("html")?.classList.add("scrollbarBeGone");
+				} else {
+					store.commit("SET_FULLSCREEN", false);
+					document.querySelector("html")?.classList.remove("scrollbarBeGone");
+				}
+			});
+
+			await store.dispatch("settings/load");
+			await store.dispatch("getNewToken");
+			await setLocale(store.state.settings.locale);
+
+			// ask the server if we are logged in or not, and update the client to reflect that status.
+			let resp = await API.get("/user");
+			if (resp.data.loggedIn) {
+				let user = resp.data;
+				delete user.loggedIn;
+				store.commit("LOGIN", user);
 			}
 		});
 
-		document.addEventListener("fullscreenchange", () => {
-			if (document.fullscreenElement) {
-				this.$store.state.fullscreen = true;
-				document.querySelector("html")?.classList.add("scrollbarBeGone");
-			} else {
-				this.$store.state.fullscreen = false;
-				document.querySelector("html")?.classList.remove("scrollbarBeGone");
-			}
-		});
+		const fullscreen = computed(() => store.state.fullscreen);
 
-		await this.$store.dispatch("settings/load");
-		await this.$store.dispatch("getNewToken");
-		await this.setLocale(this.$store.state.settings.locale);
-
-		// ask the server if we are logged in or not, and update the client to reflect that status.
-		let resp = await API.get("/user");
-		if (resp.data.loggedIn) {
-			let user = resp.data;
-			delete user.loggedIn;
-			this.$store.commit("LOGIN", user);
-		}
+		return {
+			showCreateRoomForm,
+			showLogin,
+			drawer,
+			fullscreen,
+			logout,
+			setLocale,
+			cancelRoom,
+			createTempRoom,
+			logoUrl,
+			store,
+		};
 	},
 });
+
+export default App;
 </script>
 
 <style lang="scss">
@@ -272,11 +245,29 @@ export default Vue.extend({
 	opacity: 0.7;
 }
 
+.app-bar-title {
+	margin-right: 10px;
+
+	// HACK: vuetify 3 was forcing the other buttons to center themselves.
+	flex-grow: 0;
+	flex-shrink: 0;
+	flex-basis: auto;
+}
+
 .scrollbarBeGone {
 	-ms-overflow-style: none; // I think this is an old way to do this? Probably not ideal
 	scrollbar-width: none;
 	&::-webkit-scrollbar {
 		display: none;
 	}
+}
+
+.overlay-loading-create-room {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	width: 100%;
+	height: 100%;
 }
 </style>
