@@ -25,7 +25,7 @@ export default {
 	mounted() {
 		this.beginNewVideo();
 	},
-	beforeDestroy() {
+	beforeUnmount() {
 		if (this.player) {
 			this.player.dispose();
 		}
@@ -74,7 +74,7 @@ export default {
 				this.$emit("apiready");
 				this.hasEmittedApiReady = true;
 			}
-			if (process.env.NODE_ENV === "development") {
+			if (import.meta.env.NODE_ENV === "development") {
 				for (const event of [
 					"ready",
 					"loadstart",
