@@ -6,14 +6,20 @@
 		}"
 	>
 		<div class="chat-header d-flex flex-row" v-if="activated">
-			<v-btn icon x-small @click="setActivated(false)">
-				<v-icon>fas fa-chevron-down</v-icon>
+			<v-btn
+				icon
+				size="x-small"
+				@click="setActivated(false)"
+				data-cy="chat-deactivate"
+				aria-label="close chat"
+			>
+				<v-icon>fa:fas fa-chevron-down</v-icon>
 			</v-btn>
 			<h4>{{ $t("chat.title") }}</h4>
 		</div>
 		<div ref="messages" @scroll="onScroll" class="messages d-flex flex-column flex-grow-1 mt-2">
 			<div class="d-flex flex-grow-1"><!-- Spacer --></div>
-			<transition-group name="message" style="width: 100%">
+			<transition-group name="message">
 				<div class="message" v-for="(msg, index) in chatMessagePast" :key="index">
 					<div class="from">{{ msg.from.name }}</div>
 					<div class="text">
@@ -31,11 +37,16 @@
 				</div>
 			</transition-group>
 		</div>
+		<div v-if="!stickToBottom" class="to-bottom">
+			<v-btn size="x-small" icon @click="forceToBottom">
+				<v-icon>fa:fas fa-angle-double-down</v-icon>
+			</v-btn>
+		</div>
 		<Transition name="input" @after-enter="enforceStickToBottom">
 			<div class="input-box" v-if="activated">
 				<v-text-field
-					solo
-					dense
+					variant="solo"
+					density="compact"
 					single-line
 					:placeholder="$t('chat.type-here')"
 					@keydown="onInputKeyDown"
@@ -43,12 +54,21 @@
 					autocomplete="off"
 					ref="chatInput"
 					@blur="deactivateOnBlur && setActivated(false)"
+					data-cy="chat-input"
 				/>
 			</div>
 		</Transition>
 		<div class="manual-activate" v-if="!activated">
-			<v-btn icon x-small @click="setActivated(true, (manual = true))">
-				<v-icon>far fa-comment-alt</v-icon>
+			<v-btn
+				variant="text"
+				icon
+				size="x-small"
+				@click="setActivated(true, true)"
+				color="white"
+				data-cy="chat-activate"
+				aria-label="open chat"
+			>
+				<v-icon>fa:far fa-comment-alt</v-icon>
 			</v-btn>
 		</div>
 	</div>
@@ -56,76 +76,14 @@
 
 <script lang="ts">
 import ProcessedText from "@/components/ProcessedText.vue";
-import api from "@/util/api";
-import { defineComponent, onUpdated, ref, Ref, nextTick } from "@vue/composition-api";
-import type { ChatMessage } from "common/models/types";
+import { defineComponent, onUpdated, ref, Ref, nextTick, onMounted, onUnmounted } from "vue";
+import type { ChatMessage } from "ott-common/models/types";
+import { useConnection } from "@/plugins/connection";
+import { useRoomApi } from "@/util/roomapi";
+import { ServerMessageChat } from "ott-common/models/messages";
+import { useRoomKeyboardShortcuts } from "@/util/keyboard-shortcuts";
 
 const MSG_SHOW_TIMEOUT = 20000;
-
-let inputValue = ref("");
-let stickToBottom = ref(true);
-/**
- * When chat is activated, all messages are shown. and the
- * user can scroll through message history, type in chat, etc.
- * When chat is NOT activated, when messages are received,
- * they appear and fade away after `MSG_SHOW_TIMEOUT` ms.
- */
-let activated = ref(false);
-let deactivateOnBlur = ref(false);
-/**
- * All past chat messages. They are are no longer
- * shown when deactivated.
- */
-let chatMessagePast: Ref<ChatMessage[]> = ref([]);
-/**
- * All recent chat messages that are currently shown when deactivated.
- * They will fade away after `MSG_SHOW_TIMEOUT` ms, and moved into `chatMessagePast`.
- */
-let chatMessageRecent: Ref<ChatMessage[]> = ref([]);
-let messages = ref();
-let chatInput: Ref<HTMLInputElement | undefined> = ref();
-
-function focusChatInput() {
-	chatInput.value?.focus();
-}
-
-function isActivated(): boolean {
-	return activated.value;
-}
-
-async function setActivated(value: boolean, manual = false): Promise<void> {
-	activated.value = value;
-	if (value) {
-		if (manual) {
-			deactivateOnBlur.value = false;
-		} else {
-			deactivateOnBlur.value = true;
-		}
-		await nextTick();
-		focusChatInput();
-	} else {
-		chatInput.value?.blur();
-	}
-}
-
-function onChatReceived(msg: ChatMessage): void {
-	chatMessageRecent.value.push(msg);
-	setTimeout(expireChatMessage, MSG_SHOW_TIMEOUT);
-}
-
-function expireChatMessage() {
-	chatMessagePast.value.push(chatMessageRecent.value.splice(0, 1)[0]);
-}
-
-/**
- * Performs the necessary actions to enact the stickToBottom behavior.
- */
-function enforceStickToBottom() {
-	const div = messages.value as HTMLDivElement;
-	if (stickToBottom.value) {
-		div.scrollTop = div.scrollHeight;
-	}
-}
 
 const Chat = defineComponent({
 	name: "Chat",
@@ -134,11 +92,95 @@ const Chat = defineComponent({
 	},
 	emits: ["link-click"],
 	setup() {
+		let connection = useConnection();
+		let roomapi = useRoomApi(connection);
+
+		let inputValue = ref("");
+		let stickToBottom = ref(true);
+		/**
+		 * When chat is activated, all messages are shown. and the
+		 * user can scroll through message history, type in chat, etc.
+		 * When chat is NOT activated, when messages are received,
+		 * they appear and fade away after `MSG_SHOW_TIMEOUT` ms.
+		 */
+		let activated = ref(false);
+		let deactivateOnBlur = ref(false);
+		/**
+		 * All past chat messages. They are are no longer
+		 * shown when deactivated.
+		 */
+		let chatMessagePast: Ref<ChatMessage[]> = ref([]);
+		/**
+		 * All recent chat messages that are currently shown when deactivated.
+		 * They will fade away after `MSG_SHOW_TIMEOUT` ms, and moved into `chatMessagePast`.
+		 */
+		let chatMessageRecent: Ref<ChatMessage[]> = ref([]);
+		let messages = ref();
+		let chatInput: Ref<HTMLInputElement | undefined> = ref();
+
+		let shortcuts = useRoomKeyboardShortcuts();
+		onMounted(() => {
+			connection.addMessageHandler("chat", onChatReceived);
+			if (shortcuts) {
+				shortcuts.bind({ code: "KeyT" }, () => setActivated(true, false));
+			} else {
+				console.warn("No keyboard shortcuts available");
+			}
+		});
+
+		onUnmounted(() => {
+			connection.removeMessageHandler("chat", onChatReceived);
+		});
+
+		function focusChatInput() {
+			chatInput.value?.focus();
+		}
+
+		function isActivated(): boolean {
+			return activated.value;
+		}
+
+		async function setActivated(value: boolean, manual = false): Promise<void> {
+			activated.value = value;
+			if (value) {
+				if (manual) {
+					deactivateOnBlur.value = false;
+				} else {
+					deactivateOnBlur.value = true;
+				}
+				await nextTick();
+				focusChatInput();
+			} else {
+				chatInput.value?.blur();
+				forceToBottom();
+			}
+		}
+
+		function onChatReceived(msg: ServerMessageChat): void {
+			chatMessageRecent.value.push(msg);
+			setTimeout(expireChatMessage, MSG_SHOW_TIMEOUT);
+			nextTick(enforceStickToBottom);
+		}
+
+		function expireChatMessage() {
+			chatMessagePast.value.push(chatMessageRecent.value.splice(0, 1)[0]);
+		}
+
+		/**
+		 * Performs the necessary actions to enact the stickToBottom behavior.
+		 */
+		function enforceStickToBottom() {
+			const div = messages.value as HTMLDivElement;
+			if (stickToBottom.value) {
+				div.scrollTop = div.scrollHeight;
+			}
+		}
+
 		function onInputKeyDown(e: KeyboardEvent): void {
 			if (e.key === "Enter") {
 				e.preventDefault();
 				if (inputValue.value.trim() !== "") {
-					api.chat(inputValue.value);
+					roomapi.chat(inputValue.value);
 				}
 				inputValue.value = "";
 				stickToBottom.value = true;
@@ -153,6 +195,11 @@ const Chat = defineComponent({
 			const div = messages.value as HTMLDivElement;
 			const distToBottom = div.scrollHeight - div.clientHeight - div.scrollTop;
 			stickToBottom.value = distToBottom === 0;
+		}
+
+		function forceToBottom() {
+			stickToBottom.value = true;
+			enforceStickToBottom();
 		}
 
 		onUpdated(enforceStickToBottom);
@@ -172,6 +219,7 @@ const Chat = defineComponent({
 			setActivated,
 			onChatReceived,
 			enforceStickToBottom,
+			forceToBottom,
 
 			messages,
 			chatInput,
@@ -194,9 +242,10 @@ $chat-message-bg: $background-color;
 	padding: 3px;
 	transition: all 0.2 ease;
 	pointer-events: none;
+	height: 100%;
 
 	&.activated {
-		background: rgba($color: $chat-message-bg, $alpha: 0.8);
+		background: rgba(var(--v-theme-background), $alpha: 0.8);
 		pointer-events: auto;
 	}
 }
@@ -238,7 +287,7 @@ $chat-message-bg: $background-color;
 
 .message {
 	margin: 2px 0;
-	padding: 2px 0;
+	padding: 4px;
 	opacity: 0;
 	transition: all 1s ease;
 
@@ -248,7 +297,7 @@ $chat-message-bg: $background-color;
 
 	&.recent {
 		opacity: 1;
-		background: rgba($color: $chat-message-bg, $alpha: 0.6);
+		background: rgba(var(--v-theme-background), $alpha: 0.6);
 	}
 
 	.from,
@@ -256,12 +305,12 @@ $chat-message-bg: $background-color;
 		display: inline;
 		margin: 3px 5px;
 		word-wrap: break-word;
+		overflow-wrap: anywhere;
 	}
 
 	.from {
 		font-weight: bold;
 		margin-left: 0;
-		padding: 0 4px;
 	}
 
 	@media screen and (max-width: $sm-max) {
@@ -274,6 +323,17 @@ $chat-message-bg: $background-color;
 	align-self: flex-end;
 	justify-self: end;
 	pointer-events: auto;
+}
+
+.to-bottom {
+	display: flex;
+	justify-content: start;
+	width: 100%;
+	pointer-events: auto;
+	margin: 6px 0;
+	position: absolute;
+	bottom: 42px;
+	z-index: 100;
 }
 
 // Transition animation
