@@ -1,66 +1,68 @@
 <template>
 	<div class="video-queue">
-		<div class="empty-queue" v-if="$store.state.room.queue.length === 0">
-			<v-container fill-height>
-				<v-row justify="center" align="center">
+		<div class="empty-queue" v-if="store.state.room.queue.length === 0">
+			<v-container style="height: 100%">
+				<v-row justify="center" align="center" style="height: 100%">
 					<div>
 						<div class="msg">
 							{{ $t("video-queue.no-videos") }}
 						</div>
-						<v-btn x-large block @click="$emit('switchtab')">
-							<v-icon style="margin-right: 8px">fas fa-plus</v-icon>
+						<v-btn size="x-large" block @click="$emit('switchtab')">
+							<v-icon style="margin-right: 8px">fa:fas fa-plus</v-icon>
 							{{ $t("video-queue.add-video") }}
 						</v-btn>
 					</div>
 				</v-row>
 			</v-container>
 		</div>
-		<div class="queue-controls" v-if="$store.state.room.queue.length > 0">
-			<v-btn icon @click="api.shuffle()">
-				<v-icon>fas fa-random</v-icon>
+		<div class="queue-controls" v-if="store.state.room.queue.length > 0">
+			<v-btn icon @click="roomapi.shuffle()">
+				<v-icon>fa:fas fa-random</v-icon>
 			</v-btn>
 		</div>
-		<draggable
-			v-model="$store.state.room.queue"
+		<Sortable
+			:list="store.state.room.queue"
 			:move="() => granted('manage-queue.order')"
 			@end="onQueueDragDrop"
-			handle=".drag-handle"
+			:options="{ animation: 200, handle: '.drag-handle' }"
+			item-key="id"
 		>
-			<transition-group name="video-queue">
-				<VideoQueueItem
-					v-for="(itemdata, index) in $store.state.room.queue"
-					:key="itemdata.id"
-					:item="itemdata"
-					:index="index"
-				/>
-			</transition-group>
-		</draggable>
+			<template #item="{ element }">
+				<VideoQueueItem :key="element.id" :item="element" />
+			</template>
+		</Sortable>
 	</div>
 </template>
 
 <script lang="ts">
-import { defineComponent } from "@vue/composition-api";
-import draggable from "vuedraggable";
+import { defineComponent } from "vue";
 import VideoQueueItem from "@/components/VideoQueueItem.vue";
-import api from "@/util/api";
 import { granted } from "@/util/grants";
-
-function onQueueDragDrop(e: { oldIndex: number; newIndex: number }) {
-	api.queueMove(e.oldIndex, e.newIndex);
-}
+import { useStore } from "@/store";
+import { Sortable } from "sortablejs-vue3";
+import { useConnection } from "@/plugins/connection";
+import { useRoomApi } from "@/util/roomapi";
 
 const VideoQueue = defineComponent({
 	name: "VideoQueue",
 	components: {
-		draggable,
 		VideoQueueItem,
+		Sortable,
 	},
 	setup() {
+		const store = useStore();
+		const roomapi = useRoomApi(useConnection());
+
+		function onQueueDragDrop(e: { oldIndex: number; newIndex: number }) {
+			roomapi.queueMove(e.oldIndex, e.newIndex);
+		}
+
 		return {
 			onQueueDragDrop,
 
-			api,
+			roomapi,
 			granted,
+			store,
 		};
 	},
 });
@@ -78,7 +80,7 @@ export default VideoQueue;
 	height: 300px;
 
 	.msg {
-		opacity: 0.5;
+		opacity: 0.6;
 		font-size: 20px;
 	}
 }
@@ -92,8 +94,8 @@ export default VideoQueue;
 .video-queue-leave-active {
 	transition: all 0.2s;
 }
-.video-queue-enter,
-.video-queue.leave-to {
+.video-queue-enter-from,
+.video-queue-leave-to {
 	opacity: 0;
 	transform: translateX(-30px) scaleY(0);
 }
