@@ -1,45 +1,42 @@
 <template>
-	<v-menu top offset-y :disabled="!supported">
-		<template v-slot:activator="{ on, attrs }">
-			<v-btn v-bind="attrs" v-on="on" :disabled="!supported">
-				<v-icon>fas fa-closed-captioning</v-icon>
-			</v-btn>
-		</template>
-		<v-list>
-			<v-list-item link @click="setCaptionsEnabled(true)" v-if="tracks.length === 0">
-				On
-			</v-list-item>
-			<v-list-item
-				link
-				@click="setCaptionsTrack(track)"
-				v-for="(track, idx) in tracks"
-				:key="idx"
-			>
-				{{ track }}
-			</v-list-item>
-			<v-list-item link @click="setCaptionsEnabled(false)"> Off </v-list-item>
-		</v-list>
-	</v-menu>
+	<v-btn
+		variant="text"
+		icon
+		:disabled="!supported"
+		class="media-control"
+		aria-label="Closed Captions"
+	>
+		<v-icon>fa:fas fa-closed-captioning</v-icon>
+		<v-menu location="top" offset-y activator="parent" :disabled="!supported">
+			<v-list>
+				<v-list-item link @click="setCaptionsEnabled(true)" v-if="tracks.length === 0">
+					On
+				</v-list-item>
+				<v-list-item
+					link
+					@click="setCaptionsTrack(track)"
+					v-for="(track, idx) in tracks"
+					:key="idx"
+				>
+					{{ track }}
+				</v-list-item>
+				<v-list-item link @click="setCaptionsEnabled(false)"> Off </v-list-item>
+			</v-list>
+		</v-menu>
+	</v-btn>
 </template>
 
 <script lang="ts">
-import { defineComponent, toRefs } from "@vue/composition-api";
-
-interface ClosedCaptionsSwitcherProps {
-	supported: boolean;
-	tracks: string[];
-}
+import { defineComponent, PropType } from "vue";
 
 const ClosedCaptionsSwitcher = defineComponent({
 	name: "ClosedCaptionsSwitcher",
 	emits: ["enable-cc", "cc-track"],
 	props: {
 		supported: { type: Boolean, default: true },
-		tracks: { type: Array, default: [] },
+		tracks: { type: Array as PropType<string[]>, default: () => [] },
 	},
-	setup(props: ClosedCaptionsSwitcherProps, { emit }) {
-		let { tracks } = toRefs(props);
-
+	setup(props, { emit }) {
 		function setCaptionsEnabled(value: boolean) {
 			emit("enable-cc", value);
 		}
@@ -57,3 +54,7 @@ const ClosedCaptionsSwitcher = defineComponent({
 
 export default ClosedCaptionsSwitcher;
 </script>
+
+<style lang="scss">
+@use "./media-controls.scss";
+</style>
