@@ -5,7 +5,7 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, onMounted, ref, toRefs, watch } from "@vue/composition-api";
+import { defineComponent, onMounted, ref, toRefs, watch } from "vue";
 import { getSdk } from "@/util/playerHelper.js";
 
 const DAILYMOTION_SDK_URL = "https://api.dmcdn.net/all.js";
@@ -51,7 +51,7 @@ const DailymotionPlayer = defineComponent({
 
 		onMounted(async () => {
 			let _DM = await getSdk(DAILYMOTION_SDK_URL, "DM", "dmAsyncInit");
-			DM = _DM;
+			DM.value = _DM;
 			DM.value.init({
 				status: false,
 				cookie: false,
