@@ -1,192 +1,372 @@
 <template>
-	<div>
+	<Suspense>
 		<YoutubePlayer
-			v-if="source.service == 'youtube'"
-			ref="youtube"
+			v-if="!!source && source.service == 'youtube'"
+			ref="player"
 			:video-id="source.id"
 			class="player"
-			@apiready="$emit('apiready')"
-			@playing="$emit('playing')"
-			@paused="$emit('paused')"
-			@ready="$emit('ready')"
-			@buffering="$emit('buffering')"
-			@error="$emit('error')"
+			@apiready="onApiReady"
+			@playing="onPlaying"
+			@paused="onPaused"
+			@ready="onReady"
+			@buffering="onBuffering"
+			@error="onError"
+			@buffer-progress="onBufferProgress"
 		/>
 		<VimeoPlayer
-			v-else-if="source.service == 'vimeo'"
-			ref="vimeo"
+			v-else-if="!!source && source.service == 'vimeo'"
+			ref="player"
 			:video-id="source.id"
 			class="player"
-			@apiready="$emit('apiready')"
-			@playing="$emit('playing')"
-			@paused="$emit('paused')"
-			@ready="$emit('ready')"
-			@buffering="$emit('buffering')"
-			@error="$emit('error')"
+			@apiready="onApiReady"
+			@playing="onPlaying"
+			@paused="onPaused"
+			@ready="onReady"
+			@buffering="onBuffering"
+			@error="onError"
 		/>
 		<DailymotionPlayer
-			v-else-if="source.service == 'dailymotion'"
-			ref="dailymotion"
+			v-else-if="!!source && source.service == 'dailymotion'"
+			ref="player"
 			:video-id="source.id"
 			class="player"
-			@apiready="$emit('apiready')"
-			@playing="$emit('playing')"
-			@paused="$emit('paused')"
-			@ready="$emit('ready')"
-			@buffering="$emit('buffering')"
-			@error="$emit('error')"
+			@apiready="onApiReady"
+			@playing="onPlaying"
+			@paused="onPaused"
+			@ready="onReady"
+			@buffering="onBuffering"
+			@error="onError"
 		/>
 		<GoogleDrivePlayer
-			v-else-if="source.service == 'googledrive'"
-			ref="googledrive"
+			v-else-if="!!source && source.service == 'googledrive'"
+			ref="player"
 			:video-id="source.id"
 			class="player"
-			@apiready="$emit('apiready')"
-			@playing="$emit('playing')"
-			@paused="$emit('paused')"
-			@ready="$emit('ready')"
-			@buffering="$emit('buffering')"
-			@error="$emit('error')"
+			@apiready="onApiReady"
+			@playing="onPlaying"
+			@paused="onPaused"
+			@ready="onReady"
+			@buffering="onBuffering"
+			@error="onError"
 		/>
 		<DirectPlayer
-			v-else-if="source.service == 'direct'"
-			ref="direct"
+			v-else-if="!!source && source.service == 'direct'"
+			ref="player"
 			:video-url="source.id"
-			:video-mime="source.mime"
+			:video-mime="source.mime!"
 			:thumbnail="source.thumbnail"
 			class="player"
-			@apiready="$emit('apiready')"
-			@playing="$emit('playing')"
-			@paused="$emit('paused')"
-			@ready="$emit('ready')"
-			@buffering="$emit('buffering')"
-			@error="$emit('error')"
+			@apiready="onApiReady"
+			@playing="onPlaying"
+			@paused="onPaused"
+			@ready="onReady"
+			@buffering="onBuffering"
+			@error="onError"
 			@buffer-progress="onBufferProgress"
-			@buffer-spans="timespans => $emit('buffer-spans', timespans)"
+			@buffer-spans="onBufferSpans"
 		/>
 		<GenericHlsPlayer
-			v-else-if="source.service == 'reddit'"
-			ref="reddit"
+			v-else-if="!!source && source.service == 'reddit'"
+			ref="player"
 			:videoid="source.id"
-			:hls-url="source.hls_url"
+			:hls-url="source.hls_url!"
 			:thumbnail="source.thumbnail"
 			class="player"
-			@apiready="$emit('apiready')"
-			@playing="$emit('playing')"
-			@paused="$emit('paused')"
-			@ready="$emit('ready')"
-			@buffering="$emit('buffering')"
-			@error="$emit('error')"
+			@apiready="onApiReady"
+			@playing="onPlaying"
+			@paused="onPaused"
+			@ready="onReady"
+			@buffering="onBuffering"
+			@error="onError"
 			@buffer-progress="onBufferProgress"
-			@buffer-spans="timespans => $emit('buffer-spans', timespans)"
+			@buffer-spans="onBufferSpans"
 		/>
 		<GenericHlsPlayer
-			v-else-if="source.service == 'tubi'"
-			ref="tubi"
+			v-else-if="!!source && source.service == 'tubi'"
+			ref="player"
 			:videoid="source.id"
-			:hls-url="source.hls_url"
+			:hls-url="source.hls_url!"
 			:thumbnail="source.thumbnail"
 			class="player"
-			@apiready="$emit('apiready')"
-			@playing="$emit('playing')"
-			@paused="$emit('paused')"
-			@ready="$emit('ready')"
-			@buffering="$emit('buffering')"
-			@error="$emit('error')"
+			@apiready="onApiReady"
+			@playing="onPlaying"
+			@paused="onPaused"
+			@ready="onReady"
+			@buffering="onBuffering"
+			@error="onError"
 			@buffer-progress="onBufferProgress"
-			@buffer-spans="timespans => $emit('buffer-spans', timespans)"
+			@buffer-spans="onBufferSpans"
 		/>
-		<v-container v-else fluid fill-height>
-			<v-row justify="center" align="center">
-				<v-col cols="auto">
-					<h1>{{ $t("video.no-video") }}</h1>
-					<span>{{ $t("video.no-video-text") }}</span>
-				</v-col>
-			</v-row>
+		<v-container v-else fluid fill-height class="no-video">
+			<h1>{{ $t("video.no-video") }}</h1>
+			<span>{{ $t("video.no-video-text") }}</span>
 		</v-container>
-	</div>
+		<template #fallback>
+			<v-container class="no-video">
+				<v-progress-circular indeterminate />
+			</v-container>
+		</template>
+	</Suspense>
 </template>
 
-<script>
+<script lang="ts">
+import { useStore } from "@/store";
+import { PlayerStatus } from "ott-common/models/types";
+import { QueueItem } from "ott-common/models/video";
+import { defineComponent, defineAsyncComponent, PropType, ref, Ref, computed, watch } from "vue";
+
 const services = ["youtube", "vimeo", "dailymotion", "googledrive", "direct", "reddit", "tubi"];
 
-export default {
-	name: "omniplayer",
-	props: ["source"],
-	components: {
-		YoutubePlayer: () =>
-			import(/* webpackChunkName: "youtube" */ "@/components/players/YoutubePlayer.vue"),
-		VimeoPlayer: () =>
-			import(/* webpackChunkName: "vimeo" */ "@/components/players/VimeoPlayer.vue"),
-		DailymotionPlayer: () =>
-			import(
-				/* webpackChunkName: "dailymotion" */ "@/components/players/DailymotionPlayer.vue"
-			),
-		GoogleDrivePlayer: () =>
-			import(
-				/* webpackChunkName: "googledrive" */ "@/components/players/GoogleDrivePlayer.vue"
-			),
-		DirectPlayer: () =>
-			import(/* webpackChunkName: "direct" */ "@/components/players/DirectPlayer.vue"),
-		GenericHlsPlayer: () =>
-			import(/* webpackChunkName: "hls" */ "@/components/players/GenericHlsPlayer.vue"),
+interface MediaPlayer {
+	play(): void;
+	pause(): void;
+	setVolume(volume: number): void;
+	getPosition(): number;
+	setPosition(position: number): void;
+}
+
+interface MediaPlayerWithCaptions extends MediaPlayer {
+	isCaptionsSupported(): boolean;
+	isCaptionsEnabled(): boolean;
+	setCaptionsEnabled(enabled: boolean): void;
+	getCaptionsTracks(): string[];
+	setCaptionsTrack(track: string): void;
+}
+
+export default defineComponent({
+	name: "OmniPlayer",
+	props: {
+		source: {
+			type: Object as PropType<QueueItem | null>,
+			validator: (source: QueueItem | null) => {
+				return !source || services.includes(source.service);
+			},
+		},
 	},
-	methods: {
-		player() {
-			// This can't be a computed property because of a race condition. see #355
-			if (services.includes(this.source.service)) {
-				return this.$refs[this.source.service];
+	emits: ["apiready", "playing", "paused", "ready", "buffering", "error"],
+	components: {
+		YoutubePlayer: defineAsyncComponent(() => import("./YoutubePlayer.vue")),
+		VimeoPlayer: defineAsyncComponent(() => import("./VimeoPlayer.vue")),
+		DailymotionPlayer: defineAsyncComponent(() => import("./DailymotionPlayer.vue")),
+		GoogleDrivePlayer: defineAsyncComponent(() => import("./GoogleDrivePlayer.vue")),
+		DirectPlayer: defineAsyncComponent(() => import("./DirectPlayer.vue")),
+		GenericHlsPlayer: defineAsyncComponent(() => import("./GenericHlsPlayer.vue")),
+	},
+	setup(props, { emit }) {
+		const store = useStore();
+
+		let player: Ref<MediaPlayer | null> = ref(null);
+		watch(player, () => {
+			console.debug("Player changed", player.value);
+			if (player.value) {
+				player.value.setVolume(store.state.settings.volume);
+			}
+		});
+
+		function checkForPlayer(p: MediaPlayer | null): p is MediaPlayer {
+			if (!p) {
+				console.warn(
+					`There is no player available. Is the source set? ${
+						props.source !== null
+					} Is there a player implemented for ${props.source?.service}?`
+				);
+			}
+			return !!p;
+		}
+
+		function implementsCaptions(p: MediaPlayer | null): p is MediaPlayerWithCaptions {
+			return !!p && (p as MediaPlayerWithCaptions).isCaptionsSupported !== undefined;
+		}
+
+		const isPlayerPresent = computed(() => !!player.value);
+
+		function play() {
+			if (!checkForPlayer(player.value)) {
+				return;
+			}
+			player.value.play();
+		}
+		function pause() {
+			if (!checkForPlayer(player.value)) {
+				return;
+			}
+			return player.value.pause();
+		}
+		function setVolume(volume: number) {
+			if (!checkForPlayer(player.value)) {
+				return;
+			}
+			return player.value.setVolume(volume);
+		}
+		function getPosition() {
+			if (!checkForPlayer(player.value)) {
+				return 0;
+			}
+			return player.value.getPosition();
+		}
+		function setPosition(position: number) {
+			if (!checkForPlayer(player.value)) {
+				return;
+			}
+			return player.value.setPosition(position);
+		}
+		function isCaptionsSupported() {
+			if (!checkForPlayer(player.value)) {
+				return false;
+			}
+			if (!implementsCaptions(player.value)) {
+				return false;
+			}
+			return player.value.isCaptionsSupported() ?? false;
+		}
+		function isCaptionsEnabled() {
+			if (!checkForPlayer(player.value)) {
+				return;
+			}
+			if (!implementsCaptions(player.value)) {
+				return false;
+			}
+			return player.value.isCaptionsEnabled();
+		}
+		function setCaptionsEnabled(enabled: boolean) {
+			if (!checkForPlayer(player.value)) {
+				return;
+			}
+			if (!implementsCaptions(player.value)) {
+				return false;
+			}
+			player.value.setCaptionsEnabled(enabled);
+		}
+		function toggleCaptions() {
+			setCaptionsEnabled(!isCaptionsEnabled());
+		}
+		function getCaptionsTracks(): string[] {
+			if (!checkForPlayer(player.value)) {
+				return [];
+			}
+			if (!implementsCaptions(player.value)) {
+				return [];
+			}
+			return player.value.getCaptionsTracks();
+		}
+		function setCaptionsTrack(track: string) {
+			if (!checkForPlayer(player.value)) {
+				return;
+			}
+			if (!implementsCaptions(player.value)) {
+				return;
 			}
 
-			return null;
-		},
-		play() {
-			return this.player()?.play();
-		},
-		pause() {
-			return this.player()?.pause();
-		},
-		setVolume(volume) {
-			return this.player()?.setVolume(volume);
-		},
-		getPosition() {
-			return this.player()?.getPosition();
-		},
-		setPosition(position) {
-			return this.player()?.setPosition(position);
-		},
-		onBufferProgress(percent) {
-			this.$store.commit("PLAYBACK_BUFFER", percent);
-		},
-		isCaptionsSupported() {
-			if (this.player()?.isCaptionsSupported) {
-				return this.player()?.isCaptionsSupported() ?? false;
+			if (!isCaptionsEnabled()) {
+				setCaptionsEnabled(true);
 			}
-			return false;
-		},
-		isCaptionsEnabled() {
-			return this.player()?.isCaptionsEnabled();
-		},
-		setCaptionsEnabled(value) {
-			this.player()?.setCaptionsEnabled(value);
-		},
-		toggleCaptions() {
-			this.setCaptionsEnabled(!this.isCaptionsEnabled());
-		},
-		getCaptionsTracks() {
-			this.player()?.getCaptionsTracks();
-		},
-		setCaptionsTrack(track) {
-			if (!this.isCaptionsEnabled()) {
-				this.setCaptionsEnabled(true);
+			player.value.setCaptionsTrack(track);
+		}
+
+		// player events re-emitted or data stored
+		function onApiReady() {
+			emit("apiready");
+		}
+
+		function onReady() {
+			store.commit("PLAYBACK_STATUS", PlayerStatus.ready);
+			emit("ready");
+		}
+
+		function hackReadyEdgeCase() {
+			if (
+				props.source &&
+				(props.source.service === "youtube" || props.source.service === "dailymotion")
+			) {
+				store.commit("PLAYBACK_STATUS", PlayerStatus.ready);
 			}
-			this.player()?.setCaptionsTrack(track);
-		},
+		}
+
+		function onPlaying() {
+			hackReadyEdgeCase();
+			emit("playing");
+		}
+
+		function onPaused() {
+			hackReadyEdgeCase();
+			emit("paused");
+		}
+
+		function onBuffering() {
+			store.commit("PLAYBACK_STATUS", PlayerStatus.buffering);
+			emit("buffering");
+		}
+
+		function onError() {
+			store.commit("PLAYBACK_STATUS", PlayerStatus.error);
+			emit("error");
+		}
+
+		function onBufferProgress(percent: number) {
+			store.commit("PLAYBACK_BUFFER", percent);
+		}
+
+		function onBufferSpans(spans: TimeRanges) {
+			store.commit("PLAYBACK_BUFFER_SPANS", spans);
+		}
+
+		return {
+			player,
+
+			onApiReady,
+			onReady,
+			onPlaying,
+			onPaused,
+			onBuffering,
+			onError,
+			onBufferProgress,
+			onBufferSpans,
+
+			isPlayerPresent,
+			play,
+			pause,
+			setVolume,
+			getPosition,
+			setPosition,
+			isCaptionsSupported,
+			isCaptionsEnabled,
+			setCaptionsEnabled,
+			toggleCaptions,
+			getCaptionsTracks,
+			setCaptionsTrack,
+		};
 	},
-};
+});
 </script>
 
 <style lang="scss" scoped>
+.no-video {
+	display: flex;
+	height: 100%;
+	align-items: center;
+	flex-direction: column;
+	justify-content: center;
+
+	opacity: 60%;
+	border-radius: 3px;
+}
+
+.v-theme--dark,
+.v-theme--deepblue,
+.v-theme--deepred {
+	.no-video {
+		color: #fff;
+		border: 1px solid rgba(255, 255, 255, 0.5);
+	}
+}
+
+.v-theme--light {
+	.no-video {
+		color: #000;
+		border: 1px solid rgba(0, 0, 0, 0.5);
+	}
+}
+
 .player {
 	width: 100%;
 	height: 100%;
