@@ -1,20 +1,23 @@
-import Vue from "vue";
-import VueI18n, { LocaleMessages } from "vue-i18n";
+import { nextTick } from "vue";
 import messages from "@/locales/en";
 import axios from "axios";
+import { createI18n } from "vue-i18n";
 
-Vue.use(VueI18n);
-
-export const i18n = new VueI18n({
-	locale: process.env.VUE_APP_I18N_LOCALE || "en",
-	fallbackLocale: process.env.VUE_APP_I18N_FALLBACK_LOCALE || "en",
+export const i18n = createI18n({
+	allowComposition: true,
+	locale: import.meta.env.VUE_APP_I18N_LOCALE || "en",
+	fallbackLocale: import.meta.env.VUE_APP_I18N_FALLBACK_LOCALE || "en",
 });
-i18n.setLocaleMessage("en", messages);
+i18n.global.setLocaleMessage("en", messages);
 
 const loadedLanguages = ["en"]; // our default language that is preloaded
 
 function setI18nLanguage(lang: string) {
-	i18n.locale = lang;
+	if (i18n.mode === "legacy") {
+		i18n.global.locale = lang;
+	} else {
+		i18n.global.locale.value = lang;
+	}
 	axios.defaults.headers.common["Accept-Language"] = lang;
 	document.querySelector("html")?.setAttribute("lang", lang);
 	return lang;
@@ -22,7 +25,7 @@ function setI18nLanguage(lang: string) {
 
 export async function loadLanguageAsync(lang: string): Promise<string> {
 	// If the same language
-	if (i18n.locale === lang) {
+	if (i18n.global.locale === lang) {
 		return setI18nLanguage(lang);
 	}
 
@@ -32,8 +35,10 @@ export async function loadLanguageAsync(lang: string): Promise<string> {
 	}
 
 	// If the language hasn't been loaded yet
-	const messages = await import(/* webpackChunkName: "lang-[request]" */ `@/locales/${lang}.ts`);
-	i18n.setLocaleMessage(lang, messages.default);
+	const messages = await import(`@/locales/${lang}.ts`);
+	i18n.global.setLocaleMessage(lang, messages.default);
 	loadedLanguages.push(lang);
-	return setI18nLanguage(lang);
+	setI18nLanguage(lang);
+	await nextTick();
+	return lang;
 }
