@@ -5,10 +5,11 @@ import {
 	OttApiResponseRoomGenerate,
 	OttApiResponseRoomCreate,
 	OttApiRequestRoomCreate,
-} from "common/models/rest-api";
+} from "ott-common/models/rest-api";
 import type { Store } from "vuex";
-import { Ref, ref } from "@vue/composition-api";
+import { reactive, Ref, ref } from "vue";
 import { AxiosResponse } from "axios";
+import type { FullOTTStoreState } from "@/store";
 
 /** Generate a temporary room. */
 export async function generateRoom(): Promise<OttApiResponseRoomGenerate> {
@@ -40,46 +41,33 @@ export async function createRoom(
 	}
 }
 
-export interface RoomCreateState {
-	isLoadingCreateRoom: boolean;
-	cancelledRoomCreation: boolean;
-}
-
-export let createRoomState = ref({
-	isLoadingCreateRoom: false,
-	cancelledRoomCreation: false,
-});
-
 /** Helper function to generate a temporary room, and then trigger a page navigation. */
-export async function createRoomHelper(store: Store<unknown>, options?: OttApiRequestRoomCreate) {
-	let state = createRoomState;
-	state.value.cancelledRoomCreation = false;
-	state.value.isLoadingCreateRoom = true;
+export async function createRoomHelper(
+	store: Store<FullOTTStoreState>,
+	options?: OttApiRequestRoomCreate
+) {
+	store.commit("misc/CREATING_ROOM");
 	try {
 		if (options) {
 			await createRoom(options);
-			if (state.value.cancelledRoomCreation) {
-				state.value.isLoadingCreateRoom = false;
+			if (store.state.misc.cancelledRoomCreation) {
 				return;
 			}
 			store.commit("misc/ROOM_CREATED", { name: options.name });
-			state.value.isLoadingCreateRoom = false;
 			return options.name;
 		} else {
 			let resp = await generateRoom();
-			if (state.value.cancelledRoomCreation) {
-				state.value.isLoadingCreateRoom = false;
+			if (store.state.misc.cancelledRoomCreation) {
 				return;
 			}
 			store.commit("misc/ROOM_CREATED", { name: resp.room });
-			state.value.isLoadingCreateRoom = false;
 			return resp.room;
 		}
 	} catch (err) {
-		state.value.isLoadingCreateRoom = false;
-		if (state.value.cancelledRoomCreation) {
+		if (store.state.misc.cancelledRoomCreation) {
 			return;
 		}
+		store.commit("misc/ROOM_CREATE_FAILED");
 		console.error(err);
 		store.commit("toast/ADD_TOAST", {
 			style: ToastStyle.Error,
